requirements.txt
pillow
numpy
st7789
